requirements.txt
bcrypt==4.1.2
boto3==1.18.44
botocore==1.21.65
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cryptography==42.0.5
cycler==0.12.1
filelock==3.13.4
Flask==2.0.2
Flask-Cors==3.0.10
fsspec==2024.3.1
greenlet==3.0.3
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.4.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==1.3.3
paramiko==2.8.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
s3transfer==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
shap==0.44.1
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.29
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing-extensions==4.11.0
urllib3==1.26.18
werkzeug==3.0.2
zipp==3.18.1
